+++ tests/changelog/test_changelog_update.py
@@ -15,6 +15,7 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
 from datetime import date
 from pathlib import Path
 import unittest
@@ -29,16 +30,16 @@ class ChangelogUpdateTestCase(unittest.TestCase):
         path = test_directory + 'unreleased_changelog_example.md'
         expeced_changelog = Path(
             test_directory + 'expected_changelog.md'
-        ).read_text()
+        ).read_text(encoding='utf-8')
         expeced_release_notes = Path(
             test_directory + 'expected_release_notes.md'
-        ).read_text()
+        ).read_text(encoding='utf-8')
         expeced_changelog = expeced_changelog.format(date.today().isoformat())
         expeced_release_notes = expeced_release_notes.format(
             date.today().isoformat()
         )
 
-        test_md = Path(path).read_text()
+        test_md = Path(path).read_text(encoding='utf-8')
         updated, release_notes = changelog.update(test_md, '1.2.3', 'hidden')
         self.assertEqual(expeced_changelog, updated)
         self.assertEqual(expeced_release_notes, release_notes)
@@ -59,17 +60,15 @@ something, somehing
         self.assertIs('', release_notes)
 
     def test_update_markdown_return_changelog_when_only_unreleased(self):
-        released = """
-## [1.2.3] - {}
+        released = f"""
+## [1.2.3] - {date.today().isoformat()}
 ### fixed
 so much
 ### added
 so little
 ### changed
 I don't recognize it anymore
-[1.2.3]: https://github.com/greenbone/pontos/compare/v1.0.0...v1.2.3""".format(
-            date.today().isoformat()
-        )
+[1.2.3]: https://github.com/greenbone/pontos/compare/v1.0.0...v1.2.3"""
 
         unreleased = """
 ## [Unreleased]
@@ -89,7 +88,7 @@ I don't recognize it anymore
         self.assertEqual(released.strip(), release_notes.strip())
 
     def test_update_markdown_different_unreleased_sections(self):
-        released_md = """
+        released_md = f"""
 ## [1.2.4] (Unreleased)
 ### fixed
 not so much
@@ -100,7 +99,7 @@ I don't recognize it anymore
 ### security
 [Unreleased]: https://github.com/greenbone/pontos/compare/v1.0.0...master
 
-## [1.2.3] - {}
+## [1.2.3] - {date.today().isoformat()}
 ### fixed
 so much wow
 ### added
@@ -108,9 +107,7 @@ so little
 ### changed
 I don't recognize it anymore
 [1.2.3]: https://github.com/greenbone/pontos/compare/v1.0.0...v1.2.3
-""".format(
-            date.today().isoformat()
-        )
+"""
 
         unreleased = """
 ## [1.2.4] (Unreleased)
@@ -134,8 +131,8 @@ I don't recognize it anymore
 [Unreleased]: https://github.com/greenbone/pontos/compare/v1.0.0...master
 """
 
-        released = """
-## [1.2.3] - {}
+        released = f"""
+## [1.2.3] - {date.today().isoformat()}
 ### fixed
 so much wow
 ### added
@@ -143,9 +140,7 @@ so little
 ### changed
 I don't recognize it anymore
 [1.2.3]: https://github.com/greenbone/pontos/compare/v1.0.0...v1.2.3
-""".format(
-            date.today().isoformat()
-        )
+"""
 
         test_md = unreleased
         updated, release_notes = changelog.update(
@@ -157,17 +152,15 @@ I don't recognize it anymore
         self.assertEqual(released.strip(), release_notes.strip())
 
     def test_update_markdown_return_changelog(self):
-        released = """
-## [1.2.3] - {}
+        released = f"""
+## [1.2.3] - {date.today().isoformat()}
 ### fixed
 so much
 ### added
 so little
 ### changed
 I don't recognize it anymore
-[1.2.3]: https://github.com/greenbone/pontos/compare/v1.0.0...v1.2.3""".format(
-            date.today().isoformat()
-        )
+[1.2.3]: https://github.com/greenbone/pontos/compare/v1.0.0...v1.2.3"""
 
         unreleased = """
 ## [Unreleased]
@@ -196,8 +189,8 @@ something, somehing
         self.assertEqual(released.strip(), release_notes.strip())
 
     def test_identify_heading_abort_condition_correctly(self):
-        released = """
-## [1.2.3] - {}
+        released = f"""
+## [1.2.3] - {date.today().isoformat()}
 ### fixed
 so much
 ### added
@@ -205,9 +198,7 @@ so little
 * Added support for GMP 20.08 [#254](https://github.com/greenbone/python-gvm/pull/254)
 ### changed
 I don't recognize it anymore
-[1.2.3]: https://github.com/greenbone/pontos/compare/v1.0.0...v1.2.3""".format(
-            date.today().isoformat()
-        )
+[1.2.3]: https://github.com/greenbone/pontos/compare/v1.0.0...v1.2.3"""
 
         unreleased = """
 ## [Unreleased]
@@ -248,8 +239,8 @@ something, somehing
 [Unreleased]: https://github.com/greenbone/hidden/compare/v1.2.3...HEAD
 
 """
-        released = """
-## [1.2.3] - {}
+        released = f"""
+## [1.2.3] - {date.today().isoformat()}
 ### fixed
 so much
 ### added
@@ -257,9 +248,7 @@ so little
 ### changed
 I don't recognize it anymore
 ### security
-[1.2.3]: https://github.com/greenbone/pontos/compare/v1.0.0...v1.2.3""".format(
-            date.today().isoformat()
-        )
+[1.2.3]: https://github.com/greenbone/pontos/compare/v1.0.0...v1.2.3"""
 
         test_md_template = """# Changelog
 something, somehing